requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
imbalanced-learn>=0.9.0
xgboost>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
opencv-python>=4.5.0
PyYAML>=6.0
joblib>=1.1.0